+++ lib/pages/principal.dart
@@ -2,7 +2,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
-import 'package:testemobileskore/bloc/aula_bloc.dart';
+import 'package:testemobileskore/bloc/aula.dart';
 import 'package:testemobileskore/model/aula.dart';
 import 'package:testemobileskore/repository/aula.dart';
